requirements.txt
Flask>=1.1.2
python-dotenv>=0.15.0
mariadb>=1.0.6
pycryptodome>=3.10.1